CSE535 : Mobile Computing
Assignment 01 : Health Monitoring UI

Details : Develop an UI similar to the Matlab UI. Use the emulator to display the UI. For graphing use the graphView class provided in the course web page. Take a sample array of floats. Whenever the Run button is hit the array is plotted in the canvass. On clicking the stop button the view is cleared.

Deliverables: a) The source code of your app (acknowledge reused code – mark what is your own code).  b) run-time screenshots (add a unique feature/text that distinguishes it from other screenshots, e.g. from other group’s screenshots or from the screenshots above)